refactor(tui): address clippy warnings across tui module
- Collapse nested if statements with let-chains (&&)
- Use strip_prefix instead of slice indexing where applicable
- Add Default impl for MarkdownParseBuffer
- Remove unnecessary type casts and closures
- Rename to_bytes to into_bytes for clarity
- Use iterators instead of range-based index loops
- Replace manual div_ceil with std method

+++ src/tui/markdown.rs
@@ -224,7 +224,7 @@ impl<'a> Widget for MarkdownBlockWidget<'a> {
                             if prefix_x >= max_x {
                                 break;
                             }
-                            let w = char_width(ch) as u16;
+                            let w = char_width(ch);
                             let end_x = (prefix_x + w).min(max_x);
                             for cx in prefix_x..end_x {
                                 let cell = buf.get_mut(cx, row);
@@ -270,7 +270,7 @@ impl<'a> Widget for MarkdownBlockWidget<'a> {
                             if cur_x >= max_x {
                                 break;
                             }
-                            let w = char_width(ch) as u16;
+                            let w = char_width(ch);
                             let end_x = (cur_x + w).min(max_x);
                             for cx in cur_x..end_x {
                                 let cell = buf.get_mut(cx, y);
